+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.1.7-beta.2.240617
 ### 2024/6/17
 - add aliyun fc start code
+- add ini generate to all platform
 
 ## v1.1.7-beta.1.240612
 ### 2024/6/12