+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add aliyun fc start code
 - add ini generate to all platform
 - add sis model image java nsis sh
+- disabled model image error exit
 
 ## v1.1.7-beta.1.240612
 ### 2024/6/12